WOS-200 Improve icon mapping mechanism Remove useless icons v3
