+++ application/backend/src/main/java/pl/ds/bulma/components/models/TabComponent.java
@@ -18,16 +18,12 @@ package pl.ds.bulma.components.models;
 
 import static org.apache.sling.models.annotations.DefaultInjectionStrategy.OPTIONAL;
 
-import javax.annotation.PostConstruct;
 import javax.inject.Inject;
 import lombok.Getter;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.sling.api.resource.Resource;
-import org.apache.sling.api.resource.ValueMap;
 import org.apache.sling.models.annotations.Default;
 import org.apache.sling.models.annotations.Model;
-import org.apache.sling.models.annotations.injectorspecific.SlingObject;
-import pl.ds.bulma.components.helpers.IconContainerService;
 
 @Model(adaptables = Resource.class, defaultInjectionStrategy = OPTIONAL)
 public class TabComponent {
@@ -46,30 +42,4 @@ public class TabComponent {
   @Default(values = "mdi-home-outline")
   private String icon;
 
-  @Inject
-  @Getter
-  @Default(values = "mdi")
-  private String iconLibType;
-
-  @Inject
-  @Getter
-  @Default(values = "mdi-36px")
-  private String iconSize;
-
-  @Getter
-  private String containerSize;
-
-  @SlingObject
-  private Resource resource;
-
-
-  @PostConstruct
-  private void init() {
-    IconContainerService iconContainerService = new IconContainerService(this.resource);
-    String mappingPath = "bulma/components/common/icon/containersize/defaultsizemappings";
-
-    this.containerSize
-            = iconContainerService.calculateContainerSize(this.iconLibType,
-            mappingPath, this.iconSize);
-  }
 }
