+++ application/backend/src/main/resources/libs/bulma/components/button/button.html
@@ -20,13 +20,7 @@
         <sly data-sly-test.disabled="${model.isDisabled}">
             <button class="button ${model.buttonClasses @ join=' '}" disabled>
                 <sly data-sly-use.template="./template.html"
-                     data-sly-call="${template.buttonContent @
-             label=model.label, showIconLeft=model.showIconLeft,
-             iconLibTypeLeft=model.iconLibTypeLeft, iconLeft=model.iconLeft,
-             iconSizeLeft=model.iconSizeLeft, iconContainerSizeLeft=model.iconContainerSizeLeft,
-             showIconRight=model.showIconRight, iconLibTypeRight=model.iconLibTypeRight,
-             iconRight=model.iconRight, iconSizeRight=model.iconSizeRight,
-             iconContainerSizeRight=model.iconContainerSizeRight}"></sly>
+                     data-sly-call="${template.buttonContent @ model = model}"></sly>
             </button>
         </sly>
         <sly data-sly-test="${!disabled}">
@@ -34,26 +28,14 @@
                 <button class="button js-modal-trigger ${model.buttonClasses @ join=' '}"
                         data-target="${model.modalId}">
                     <sly data-sly-use.template="./template.html"
-                         data-sly-call="${template.buttonContent @
-             label=model.label, showIconLeft=model.showIconLeft,
-             iconLibTypeLeft=model.iconLibTypeLeft, iconLeft=model.iconLeft,
-             iconSizeLeft=model.iconSizeLeft, iconContainerSizeLeft=model.iconContainerSizeLeft,
-             showIconRight=model.showIconRight, iconLibTypeRight=model.iconLibTypeRight,
-             iconRight=model.iconRight, iconSizeRight=model.iconSizeRight,
-             iconContainerSizeRight=model.iconContainerSizeRight}"></sly>
+                         data-sly-call="${template.buttonContent @ model = model}"></sly>
                 </button>
             </sly>
 
             <sly data-sly-test="${model.actionType != 'open-modal'}">
                 <button class="button ${model.buttonClasses @ join=' '}">
                     <sly data-sly-use.template="./template.html"
-                         data-sly-call="${template.buttonContent @
-             label=model.label, showIconLeft=model.showIconLeft,
-             iconLibTypeLeft=model.iconLibTypeLeft, iconLeft=model.iconLeft,
-             iconSizeLeft=model.iconSizeLeft, iconContainerSizeLeft=model.iconContainerSizeLeft,
-             showIconRight=model.showIconRight, iconLibTypeRight=model.iconLibTypeRight,
-             iconRight=model.iconRight, iconSizeRight=model.iconSizeRight,
-             iconContainerSizeRight=model.iconContainerSizeRight}"></sly>
+                         data-sly-call="${template.buttonContent @ model = model}"></sly>
                 </button>
             </sly>
         </sly>
@@ -62,13 +44,7 @@
         <a href="${model.url}" class="button ${model.buttonClasses @ join=' '}"
            target="${model.openInNewTab && model.url != '' ? '_blank' : ''}">
             <sly data-sly-use.template="./template.html"
-                 data-sly-call="${template.buttonContent @
-             label=model.label, showIconLeft=model.showIconLeft,
-             iconLibTypeLeft=model.iconLibTypeLeft, iconLeft=model.iconLeft,
-             iconSizeLeft=model.iconSizeLeft, iconContainerSizeLeft=model.iconContainerSizeLeft,
-             showIconRight=model.showIconRight, iconLibTypeRight=model.iconLibTypeRight,
-             iconRight=model.iconRight, iconSizeRight=model.iconSizeRight,
-             iconContainerSizeRight=model.iconContainerSizeRight}"></sly>
+                 data-sly-call="${template.buttonContent @ model = model}"></sly>
         </a>
     </sly>
 </sly>
