+++ application/backend/src/main/resources/libs/bulma/components/icon/icon.html
@@ -25,7 +25,7 @@
       <sly data-sly-use.template="./template.html"
            data-sly-call="${template.icon @
              colorOnText=model.colorOnText, textVariant=model.textVariant,
-             icon=model.icon, addText=model.addText, label=model.label, containerSize=model.containerSize, iconSize=model.iconSize, iconLibType=model.iconLibType}"></sly>
+             icon=model.icon, addText=model.addText, label=model.label, containerSize=model.containerSize, iconSize=model.iconSize, iconLibType=model.iconLibType, hidden=model.hidden}"></sly>
     </div>
   </sly>
   <sly data-sly-test="${model.elementType == 'span'}">
@@ -34,7 +34,7 @@
       <sly data-sly-use.template="./template.html"
            data-sly-call="${template.icon @
            colorOnText=model.colorOnText, textVariant=model.textVariant,
-           icon=model.icon, addText=model.addText, label=model.label, containerSize=model.containerSize, iconSize=model.iconSize, iconLibType=model.iconLibType}"></sly>
+           icon=model.icon, addText=model.addText, label=model.label, containerSize=model.containerSize, iconSize=model.iconSize, iconLibType=model.iconLibType, hidden=model.hidden}"></sly>
     </span>
   </sly>
 </sly>
