+++ application/backend/src/main/resources/libs/bulma/components/icon/template.html
@@ -13,9 +13,9 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 */-->
-<template data-sly-template.icon="${ @ colorOnText, textVariant, icon, addText, label, containerSize, iconSize, iconLibType}">
+<template data-sly-template.icon="${ @ colorOnText, textVariant, icon, addText, label, containerSize, iconSize, iconLibType, hidden}">
     <span class="icon ${containerSize} ${!colorOnText  && textVariant ? textVariant : ''}">
-      <i class="${iconLibType} ${iconSize} ${icon}"></i>
+      <i class="${iconLibType} ${iconSize} ${icon}" aria-hidden="${hidden}"></i>
     </span>
     <sly data-sly-test="${addText}">
         <span>${label}</span>
