+++ application/backend/src/main/resources/libs/bulma/components/icon/template/.content.json
@@ -7,5 +7,6 @@
   "iconLibType": "mdi",
   "iconSize": "mdi-36px",
   "selectOrType": "select",
+  "hidden": true,
   "textVariant": ""
 }
