+++ application/backend/src/main/resources/libs/bulma/components/navbar/navbarbrand/template/.content.json
@@ -5,8 +5,12 @@
     "jcr:primaryType": "nt:unstructured",
     "sling:resourceType": "bulma/components/navbar/navbaritem",
     "label": "Brand",
-    "icon": "none",
     "type": "link",
-    "url": ""
+    "url": "",
+    "selectOrType": "select",
+    "addIcon": false,
+    "icon": "mdi-home-outline",
+    "iconLibType": "mdi",
+    "iconSize": "mdi-36px"
   }
 }
