+++ application/backend/src/main/resources/libs/bulma/components/navbar/navbarend/template/.content.json
@@ -4,6 +4,12 @@
   "item": {
     "jcr:primaryType": "nt:unstructured",
     "sling:resourceType": "bulma/components/navbar/navbaritem",
-    "label": "Item"
+    "label": "Item",
+    "type": "link",
+    "addIcon": false,
+    "selectOrType": "select",
+    "icon": "mdi-home-outline",
+    "iconLibType": "mdi",
+    "iconSize": "mdi-36px"
   }
 }
