+++ application/backend/src/main/resources/libs/bulma/components/navbar/navbaritem/dialog/.content.json
@@ -41,9 +41,102 @@
           "name": "url",
           "label": "URL connected to item"
         },
-        "icon": {
-          "sling:resourceType": "wcm/dialogs/components/include",
-          "path": "/libs/bulma/components/common/menuicon"
+        "addIcon": {
+          "sling:resourceType": "wcm/dialogs/components/toggle",
+          "name": "addIcon",
+          "label": "Add icon?"
+        },
+        "iconContainer": {
+          "sling:resourceType": "wcm/dialogs/components/container",
+          "iconLibType": {
+            "sling:resourceType": "wcm/dialogs/components/select",
+            "name": "iconLibType",
+            "label": "Icon Library",
+            "required": true,
+            "datasource": {
+              "sling:resourceType": "/apps/iconLibraryDatasource"
+            }
+          },
+          "selectOrType": {
+            "sling:resourceType": "wcm/dialogs/components/radio",
+            "name": "selectOrType",
+            "label": "Select icon or type it",
+            "required": true,
+            "select": {
+              "sling:resourceType": "wcm/dialogs/components/radio/option",
+              "label": "Select it",
+              "value": "select",
+              "selected": true
+            },
+            "type": {
+              "sling:resourceType": "wcm/dialogs/components/radio/option",
+              "label": "Type it",
+              "value": "type"
+            }
+          },
+          "typedIcon": {
+            "sling:resourceType": "wcm/dialogs/components/textfield",
+            "name": "icon",
+            "label": "Name of the icon",
+            "ws:display": {
+              "condition": {
+                "sourceName": "selectOrType",
+                "values": "type"
+              }
+            }
+          },
+          "iconFA": {
+            "sling:resourceType": "wcm/dialogs/components/container",
+            "selectableIcon": {
+              "sling:resourceType": "wcm/dialogs/components/include",
+              "path": "bulma/components/common/icon/icons/fontawesome",
+              "ws:display": {
+                "condition": {
+                  "sourceName": "selectOrType",
+                  "values": "select"
+                }
+              }
+            },
+            "iconSize": {
+              "sling:resourceType": "wcm/dialogs/components/include",
+              "path": "bulma/components/common/icon/iconsize/fontawesome"
+            },
+            "ws:display": {
+              "condition": {
+                "sourceName": "iconLibType",
+                "values": "fas"
+              }
+            }
+          },
+          "iconMD": {
+            "sling:resourceType": "wcm/dialogs/components/container",
+            "selectableIcon": {
+              "sling:resourceType": "wcm/dialogs/components/include",
+              "path": "bulma/components/common/icon/icons/materialdesign",
+              "ws:display": {
+                "condition": {
+                  "sourceName": "selectOrType",
+                  "values": "select"
+                }
+              }
+            },
+            "iconSize": {
+              "sling:resourceType": "wcm/dialogs/components/include",
+              "path": "bulma/components/common/icon/iconsize/materialdesign"
+            },
+            "ws:display": {
+              "condition": {
+                "sourceName": "iconLibType",
+                "values": "mdi"
+              }
+            }
+          },
+          "ws:display": {
+            "condition": {
+              "sourceName": "addIcon",
+              "values": "true"
+            }
+          }
         },
         "ws:display": {"condition": {
           "sourceName": "type",
