+++ application/backend/src/main/resources/libs/bulma/components/navbar/navbarstart/template/.content.json
@@ -4,7 +4,13 @@
   "item": {
     "jcr:primaryType": "nt:unstructured",
     "sling:resourceType": "bulma/components/navbar/navbaritem",
-    "label": "Item"
+    "label": "Item",
+    "type": "link",
+    "addIcon": false,
+    "selectOrType": "select",
+    "icon": "mdi-home-outline",
+    "iconLibType": "mdi",
+    "iconSize": "mdi-36px"
   },
   "dropdown": {
     "jcr:primaryType": "nt:unstructured",
