+++ application/backend/src/main/resources/libs/bulma/components/navbar/template/.content.json
@@ -10,9 +10,13 @@
       "jcr:primaryType": "nt:unstructured",
       "sling:resourceType": "bulma/components/navbar/navbaritem",
       "label": "Brand",
-      "icon": "none",
+      "url": "",
       "type": "link",
-      "url": ""
+      "addIcon": false,
+      "selectOrType": "select",
+      "icon": "mdi-home-outline",
+      "iconLibType": "mdi",
+      "iconSize": "mdi-36px"
     }
   },
   "menu": {
@@ -25,9 +29,12 @@
         "jcr:primaryType": "nt:unstructured",
         "sling:resourceType": "bulma/components/navbar/navbaritem",
         "label": "Item",
-        "icon": "none",
         "type": "link",
-        "url": ""
+        "addIcon": false,
+        "selectOrType": "select",
+        "icon": "mdi-home-outline",
+        "iconLibType": "mdi",
+        "iconSize": "mdi-36px"
       },
       "dropdown": {
         "jcr:primaryType": "nt:unstructured",
@@ -53,9 +60,12 @@
         "jcr:primaryType": "nt:unstructured",
         "sling:resourceType": "bulma/components/navbar/navbaritem",
         "label": "Item",
-        "icon": "none",
         "type": "link",
-        "url": ""
+        "addIcon": false,
+        "selectOrType": "select",
+        "icon": "mdi-home-outline",
+        "iconLibType": "mdi",
+        "iconSize": "mdi-36px"
       }
     }
   }
