+++ application/backend/src/main/resources/libs/bulma/components/panel/panelblock/panelblock.html
@@ -20,7 +20,7 @@
   <sly data-sly-test="${resource.hasChildren}">
       <a class="panel-block ${model.containerSize}" id="${model.id}">
         <sly data-sly-test="${model.addIcon && model.icon}">
-          <span class="panel-icon"><i class="${model.iconLibType} ${model.iconSize} ${model.icon}" aria-hidden="true"></i></span>
+          <sly data-sly-resource="${resource @ resourceType = 'bulma/components/icon', requestAttributes = wcmmode.disabledModeAttribute}"></sly>
         </sly>
         <sly data-sly-list="${resource.children}">
           <sly data-sly-resource="${item}"></sly>
