+++ application/backend/src/main/resources/libs/bulma/components/panel/panelblock/template/.content.json
@@ -5,5 +5,6 @@
   "icon": "mdi-home-outline",
   "iconLibType": "mdi",
   "iconSize": "mdi-36px",
+  "hidden": "true",
   "selectOrType": "select"
 }
