+++ application/backend/src/main/resources/libs/bulma/components/tabs/template/.content.json
@@ -6,6 +6,12 @@
     "jcr:primaryType": "nt:unstructured",
     "sling:resourceType": "bulma/components/tabs/tab",
     "label":"Tab 1",
+    "icon": "mdi-home-outline",
+    "addIcon": false,
+    "hidden": "true",
+    "selectOrType": "select",
+    "iconLibType": "mdi",
+    "iconSize": "mdi-36px",
     "tabcontent": {
       "jcr:primaryType": "nt:unstructured",
       "sling:resourceType": "bulma/components/tabs/tabcontent"
@@ -15,6 +21,12 @@
     "jcr:primaryType": "nt:unstructured",
     "sling:resourceType": "bulma/components/tabs/tab",
     "label":"Tab 2",
+    "icon": "mdi-home-outline",
+    "addIcon": false,
+    "hidden": "true",
+    "selectOrType": "select",
+    "iconLibType": "mdi",
+    "iconSize": "mdi-36px",
     "tabcontent": {
       "jcr:primaryType": "nt:unstructured",
       "sling:resourceType": "bulma/components/tabs/tabcontent"
@@ -24,6 +36,12 @@
     "jcr:primaryType": "nt:unstructured",
     "sling:resourceType": "bulma/components/tabs/tab",
     "label":"Tab 3",
+    "icon": "mdi-home-outline",
+    "addIcon": false,
+    "hidden": "true",
+    "selectOrType": "select",
+    "iconLibType": "mdi",
+    "iconSize": "mdi-36px",
     "tabcontent": {
       "jcr:primaryType": "nt:unstructured",
       "sling:resourceType": "bulma/components/tabs/tabcontent"
